Fix: Remove redundant (int)strlen casts and use size_t for string lengths
- Removed all (int)strlen casts across 6 dbgen source files (35+ instances)
- Changed struct fields from int to size_t: alen, clen, tlen, nlen
- Updated dsstypes.h and tpch_dbgen.h to use proper size_t type
- Eliminated all strlen-related conversion warnings
- Added #include <stddef.h> to tpch_dbgen.h for size_t definition
- Follows C standards: strlen() returns size_t, not int

+++ dbgen/bm_utils.c
@@ -355,7 +355,7 @@ long      weight,
             continue;
             }
         target->list[count].text =
-            (char *) malloc((size_t)((int)strlen(token) + 1));
+            (char *) malloc(strlen(token) + 1);
         MALLOC_CHECK(target->list[count].text);
         strcpy(target->list[count].text, token);
         target->max += weight;
@@ -443,7 +443,7 @@ agg_str(distribution *set, long count, long col, char *dest)
 		strcat(dest, DIST_MEMBER(set,DIST_PERMUTE(d, i)));
 		strcat(dest, " ");
 		}
-	*(dest + (int)strlen(dest) - 1) = '\0';
+	*(dest + strlen(dest) - 1) = '\0';
 
     return;
 }
